Move icon from public folder
Reflame doesn't have a separate `public` and `src` folder. The `public` folder in Reflame _is_ the `src` folder (we default to calling this folder `app`). This means you can import modules using absolute paths like "/App.jsx", which is natively supported by browsers (though not supported by Vite, so we'll have to revert to relative imports here).

Also note that in Reflame the favicon is automatically set to the icon.svg file in the root folder.